Fix Typo In README.md
I guess you didn't copy the example from working code, lol. It's kind of disappointing when you have to debug the example code for the library you're using.

In good news, it worked great after that! My only other comment would be that the "vaultAppName" part of the config seemed to be a confusing name. The value I needed to pass there had nothing to do with my function name (which is passed as default), and instead was aligned with the name of the Vault role name I'm asking to use as the IAM user.